Multiple changes: - change arduinoFFT to float (custom) - update audioreactive to use float - update effects to use float - info slider (usermod) - hide Peek in 2D - minor bugfixes

+++ usermods/audioreactive/audio_source.h
@@ -50,7 +50,7 @@ class AudioSource {
        Read num_samples from the microphone, and store them in the provided
        buffer
     */
-    virtual void getSamples(double *buffer, uint16_t num_samples) = 0;
+    virtual void getSamples(float *buffer, uint16_t num_samples) = 0;
 
     /* Get an up-to-date sample without DC offset */
     virtual int getSampleWithoutDCOffset() { return _sampleNoDCOffset; };
@@ -156,7 +156,7 @@ class I2SSource : public AudioSource {
       if (_mclkPin != I2S_PIN_NO_CHANGE) pinManager.deallocatePin(_mclkPin, PinOwner::UM_Audioreactive);
     }
 
-    virtual void getSamples(double *buffer, uint16_t num_samples) {
+    virtual void getSamples(float *buffer, uint16_t num_samples) {
       if (_initialized) {
         esp_err_t err;
         size_t bytes_read = 0;        /* Counter variable to check if we actually got enough data */
@@ -184,17 +184,17 @@ class I2SSource : public AudioSource {
           if (_shift != 0)
               newSamples[i] >>= 16;
 #endif
-          double currSample = 0.0;
+          float currSample = 0.0f;
           if(_shift > 0)
-            currSample = (double) (newSamples[i] >> _shift);
+            currSample = (float) (newSamples[i] >> _shift);
           else {
             if(_shift < 0)
-              currSample = (double) (newSamples[i] << (- _shift)); // need to "pump up" 12bit ADC to full 16bit as delivered by other digital mics
+              currSample = (float) (newSamples[i] << (- _shift)); // need to "pump up" 12bit ADC to full 16bit as delivered by other digital mics
             else
 #ifdef I2S_SAMPLE_DOWNSCALE_TO_16BIT
-              currSample = (double) newSamples[i] / 65536.0;        // _shift == 0 -> use the chance to keep lower 16bits
+              currSample = (float) newSamples[i] / 65536.0f;      // _shift == 0 -> use the chance to keep lower 16bits
 #else
-              currSample = (double) newSamples[i];
+              currSample = (float) newSamples[i];
 #endif
           }
           buffer[i] = currSample;
@@ -356,7 +356,7 @@ class I2SAdcSource : public I2SSource {
       _initialized = true;
     }
 
-    void getSamples(double *buffer, uint16_t num_samples) {
+    void getSamples(float *buffer, uint16_t num_samples) {
       /* Enable ADC. This has to be enabled and disabled directly before and
        * after sampling, otherwise Wifi dies
        */
